refactor(objects): split route into objects/ subdirectory
Aligns the /connections/:name/* route file layout with the
admin/users and connections routes — a dedicated subdirectory
holding the route component, its loader, and colocated tests.

- New app/routes/objects/objects.loader.ts contains the LoaderFunctionArgs
  handler and the BucketRouteLoaderResponse type; all server-only imports
  (AWS SDK, authMiddleware helpers, getConnection, pinnedPaths.server)
  live in this file.
- New app/routes/objects/objects.route.tsx is the UI component, the
  middleware/meta/handle exports, and a re-export of loader + type.
- The test file moves into app/routes/objects/__tests__/ and its import
  path follows.
- app/routes.ts points at the new file path.

No behavioural change. Pure move and split.
